@@ -1,14 +1,17 @@
-
+package com.test.pg.secure.sampleapp;
 
 public class SampleAppConstants {
     //API_KEY is given by the Payment Gateway. Please Copy Paste Here.
-    public static final String PG_API_KEY = "XXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX";
+    public static final String PG_API_KEY = "YOUR_API_KEY";
+
+    //SALT_KEY is given by the Payment Gateway. Please Copy Paste Here.
+    public static final String PG_SALT_KEY = "YOUR_SALT_KEY";
 
     //URL to Accept Payment Response After Payment. This needs to be done at the client's web server.
-    public static final String PG_RETURN_URL = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX";
+    public static final String PG_RETURN_URL = "YOUR_RETURN_URL";
 
     //Enter the Mode of Payment Here . Allowed Values are "LIVE" or "TEST".
-    public static final String PG_MODE = "TEST";
+    public static final String PG_MODE = "LIVE";
 
     //PG_CURRENCY is given by the Payment Gateway. Only "INR" Allowed.
     public static final String PG_CURRENCY = "INR";
@@ -16,21 +19,21 @@ public class SampleAppConstants {
     //PG_COUNTRY is given by the Payment Gateway. Only "IND" Allowed.
     public static final String PG_COUNTRY = "IND";
 
-    public static final String PG_AMOUNT = "2000";
-    public static final String PG_EMAIL = "fdhaskjdfh@gmail.com";
-    public static final String PG_NAME = "test";
-    public static final String PG_PHONE = "8333873777";
+    public static final String PG_AMOUNT = "2";
+    public static final String PG_EMAIL = "XXXX";
+    public static final String PG_NAME = "XXXX";
+    public static final String PG_PHONE = "XXXXX";
     public static       String PG_ORDER_ID = "";
-    public static final String PG_DESCRIPTION = "hv;hghkjgg";
-    public static final String PG_CITY = "xvgf";
-    public static final String PG_STATE = "Andhra Pradesh";
-    public static final String PG_ADD_1 = "dsfg";
-    public static final String PG_ADD_2 = "xzdvdf";
-    public static final String PG_ZIPCODE = "522001";
-    public static final String PG_UDF1 = "";
-    public static final String PG_UDF2 = "";
-    public static final String PG_UDF3 = "";
-    public static final String PG_UDF4 = "";
-    public static final String PG_UDF5 = "";
+    public static final String PG_DESCRIPTION = "XXXX";
+    public static final String PG_CITY = "Chennai";
+    public static final String PG_STATE = "Tamilnadu";
+    public static final String PG_ADD_1 = "address1";
+    public static final String PG_ADD_2 = "address2";
+    public static final String PG_ZIPCODE = "XXXXX";
+    public static final String PG_UDF1 = "udf1";
+    public static final String PG_UDF2 = "udf2";
+    public static final String PG_UDF3 = "udf3";
+    public static final String PG_UDF4 = "udf4";
+    public static final String PG_UDF5 = "udf5";
 
 }
